refactor: 수정, 삭제 드롭다운 공통 컴포넌트로 사용할 수 있도록 개선 [TM-122]
