+++ src/shared/components/dropdown/index.ts
@@ -1,1 +1,2 @@
-export * from './compound/Dropdown';
+export * from './plate/Dropdown';
+export * from './ActionDropdown';
